+++ src/core/index.ts
@@ -3,4 +3,5 @@ export { levelHint } from "./indent/levelHint";
 export { levelFolding } from "./folding/levelFolding";
 export { validator } from "./validator/validator";
 export { semanticTokens, legend } from "./semantic";
+export * from "./utils/findNodeAtPosition";
 export * from "./types";