+++ src/server.ts
@@ -1,5 +1,5 @@
 import { createConnection, ProposedFeatures } from "vscode-languageserver/node";
-import { createServer } from "./server";
+import { createServer } from "./lsp-server";
 
 const connection = createConnection(ProposedFeatures.all);