+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.0.3
+
+- Basic validation of node values
+- Added support for Xref references and pointer definitions
+
 ### 0.0.1
 
 - Initial release